Memperbarui source kode bagian db firebase

+++ freegul_version_1/app/src/main/java/com/example/freegul_version_1/Patient.kt
@@ -4,7 +4,7 @@ data class Patient(
     val key: String = "",
     val jenis_kelamin: String = "",
     val nama_lengkap: String = "",
-    val nilai_gula_darah: Int = 0,
+    val nilai_gula_darah: Double = 0.0,
     val no_handphone: String = "",
     val tanggal: String = "",
     val usia: Int = 0,
